XIVY-12289 return selection after deletion
When calling `deleteFirstSelectedRow` the selection in the table is changed by using some logic.
The value of the new selection is usually needed after calling this function.
E.g. in the data class editor, after deleting the selected row, the selected field needs to be set to the same selection in the table.
Instead of recalculating this value using the same logic, it will be returned from `deleteFirstSelectedRow`.