XIVY-15943 export `adjustSelectionAfterDeletionOfRow`
Export the table utility method `adjustSelectionAfterDeletionOfRow` to be used in the CMS Editor.
This is useful since the data in the CMS Editor is not changed directly but refetched from the back-end.
Because of this, I don't need the additional processing of the data made in `deleteFirstSelectedRow` when deleting a Content Object.
I just want to know what row to select after the fact.